Login (Almost) Complete
Login UI is complete. A default userList is build and entered credentials are compared to the default data. If valid information is entered, the user's ID is shown in the label on the login screen. This needs to be changed so that instead of the user ID shown, that the login screen is hidden and the home screen is shown.